ai_auto_fill/requirements.txt
bs4
requests
openai
scrapy
openpyxl
python-docx
python-pptx
nltk
